requirements.txt
torch
torchvision
torchtext
torchaudio
PyYAML
librosa
pandas
transformers
TextGrid
sentencepiece
fastapi
uvicorn
protobuf
datasets
accelerate
python-multipart